removes question list component
